+++ src/PostgREST/Query/SqlFragment.hs
@@ -51,7 +51,8 @@ import Text.InterpolatedString.Perl6 (qc)
 
 import PostgREST.DbStructure.Identifiers (FieldName,
                                           QualifiedIdentifier (..))
-import PostgREST.MediaType               (MTPlanOption (..))
+import PostgREST.MediaType               (MTPlanFormat (..),
+                                          MTPlanOption (..))
 import PostgREST.RangeQuery              (NonnegRange, allRange,
                                           rangeLimit, rangeOffset)
 import PostgREST.Request.ReadQuery       (SelectItem)
@@ -370,10 +371,10 @@ intercalateSnippet :: ByteString -> [SQL.Snippet] -> SQL.Snippet
 intercalateSnippet _ [] = mempty
 intercalateSnippet frag snippets = foldr1 (\a b -> a <> SQL.sql frag <> b) snippets
 
-explainF :: [MTPlanOption] -> SQL.Snippet -> SQL.Snippet
-explainF opts snip =
+explainF :: MTPlanFormat -> [MTPlanOption] -> SQL.Snippet -> SQL.Snippet
+explainF fmt opts snip =
   "EXPLAIN (" <>
-    SQL.sql (BS.intercalate ", " ("FORMAT JSON" : (fmtPlanOpt <$> opts))) <>
+    SQL.sql (BS.intercalate ", " (fmtPlanFmt fmt : (fmtPlanOpt <$> opts))) <>
   ") " <> snip
   where
     fmtPlanOpt :: MTPlanOption -> BS.ByteString
@@ -382,3 +383,6 @@ explainF opts snip =
     fmtPlanOpt PlanSettings = "SETTINGS"
     fmtPlanOpt PlanBuffers  = "BUFFERS"
     fmtPlanOpt PlanWAL      = "WAL"
+
+    fmtPlanFmt PlanJSON = "FORMAT JSON"
+    fmtPlanFmt PlanText = "FORMAT TEXT"